fixup! fixup! fixup! Added workspace successfull start metric
