+++ wsmaster/che-core-api-metrics/src/main/java/org/eclipse/che/api/metrics/WorkspaceStartAttemptsMeterBinder.java
@@ -17,33 +17,35 @@ import com.google.inject.Inject;
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.binder.MeterBinder;
+import javax.inject.Singleton;
 import org.eclipse.che.api.core.model.workspace.WorkspaceStatus;
 import org.eclipse.che.api.core.notification.EventService;
 import org.eclipse.che.api.workspace.shared.dto.event.WorkspaceStatusEvent;
 
-/** Counts number of successfully started workspaces. */
-public class WorkspaceStartedMeterBinder implements MeterBinder {
+/** Counts number of attempts to start workspace. */
+@Singleton
+public class WorkspaceStartAttemptsMeterBinder implements MeterBinder {
   private final EventService eventService;
 
   private Counter startedCounter;
 
   @Inject
-  public WorkspaceStartedMeterBinder(EventService eventService) {
+  public WorkspaceStartAttemptsMeterBinder(EventService eventService) {
     this.eventService = eventService;
   }
 
   @Override
   public void bindTo(MeterRegistry registry) {
     startedCounter =
-        Counter.builder(workspaceMetric("started.total"))
-            .description("The count of started workspaces")
+        Counter.builder(workspaceMetric("starting_attempts.total"))
+            .description("The count of workspaces start attempts")
             .register(registry);
 
     // only subscribe to the event once we have the counters ready
     eventService.subscribe(
         event -> {
-          if (event.getPrevStatus() == WorkspaceStatus.STARTING
-              && event.getStatus() == WorkspaceStatus.RUNNING) {
+          if (event.getPrevStatus() == WorkspaceStatus.STOPPED
+              && event.getStatus() == WorkspaceStatus.STARTING) {
             startedCounter.increment();
           }
         },
