+++ wsmaster/che-core-api-metrics/src/main/java/org/eclipse/che/api/metrics/WsMasterMetricsModule.java
@@ -28,7 +28,7 @@ public class WsMasterMetricsModule extends AbstractModule {
 
     meterMultibinder.addBinding().to(WorkspaceActivityMeterBinder.class);
     meterMultibinder.addBinding().to(WorkspaceFailureMeterBinder.class);
-    meterMultibinder.addBinding().to(WorkspaceStartedMeterBinder.class);
-    meterMultibinder.addBinding().to(WorkspaceStartingAttemptsMeterBinder.class);
+    meterMultibinder.addBinding().to(WorkspaceSuccessfulStartAttemptsMeterBinder.class);
+    meterMultibinder.addBinding().to(WorkspaceStartAttemptsMeterBinder.class);
   }
 }
